change double to BigDecimal

+++ src/main/java/com/example/criptoparser/dto/mapper/CurrencyPriceMapper.java
@@ -5,11 +5,13 @@ import com.example.criptoparser.dto.CurrencyDto;
 import com.example.criptoparser.model.CurrencyPrice;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class CurrencyPriceMapper {
     public CurrencyPrice toModel(ApiCurrencyDto apiCurrencyDto) {
         CurrencyPrice currencyPrice = new CurrencyPrice();
-        currencyPrice.setPrice(apiCurrencyDto.getPrice());
+        currencyPrice.setPrice(BigDecimal.valueOf(apiCurrencyDto.getPrice()));
         currencyPrice.setCryptoCurrency(apiCurrencyDto.getCryptoCurrency());
         currencyPrice.setFiatCurrency(apiCurrencyDto.getFiatCurrency());
         return currencyPrice;
@@ -17,7 +19,7 @@ public class CurrencyPriceMapper {
 
     public CurrencyPrice toModel(CurrencyDto currencyDto) {
         CurrencyPrice currencyPrice = new CurrencyPrice();
-        currencyPrice.setPrice(currencyDto.getPrice());
+        currencyPrice.setPrice(BigDecimal.valueOf(currencyDto.getPrice()));
         currencyPrice.setCryptoCurrency(currencyDto.getCryptoCurrency());
         currencyPrice.setFiatCurrency(currencyDto.getFiatCurrency());
         return currencyPrice;
@@ -26,7 +28,7 @@ public class CurrencyPriceMapper {
     public CurrencyDto toDto(CurrencyPrice currencyPrice) {
         CurrencyDto currencyDto = new CurrencyDto();
         currencyDto.setId(currencyPrice.getId());
-        currencyDto.setPrice(currencyPrice.getPrice());
+        currencyDto.setPrice(currencyPrice.getPrice().doubleValue());
         currencyDto.setCryptoCurrency(currencyPrice.getCryptoCurrency());
         currencyDto.setFiatCurrency(currencyPrice.getFiatCurrency());
         return currencyDto;
